Code cleanup/Added Comments
Cleaned up some reused code and added comments for easier readability. Also added a confirmation message using JS's confirm() method when deleting an employee/office.

+++ Angular/project-1/src/app/employees/employees.component.ts
@@ -16,10 +16,18 @@ import { Office } from '../models/office';
 })
 export class EmployeesComponent {
 
+  // this is the array that will hold all of our employees
+  // We create an employee card for every item in the array
   employees: Employee[] = [];
+
+  // this is the searchId that will be used to search for an employee
   searchId: number = 0;
+
+  // error flag for when an employee does not exist when searching by id
   employeeDoesNotExist: boolean = false;
+  // error flag for when an office does not exist when adding an employee
   officeDoesNotExist: boolean = false;
+  // error flag for when an office is at max capacity when adding an employee
   maxEmployees: boolean = false;
 
   // this injects HttpClient for us to use
@@ -30,12 +38,17 @@ export class EmployeesComponent {
     this.getAllEmployees();
   }
 
+  // this is the object that will be bound to the form
   formEmployee: Employee = new Employee(0, "", "", "", "", 0, new Office(0, "", "", 0, []));
+
+  // this method will be called when the form is submitted
   addEmployee() {
     this.httpService.createEmployee(this.formEmployee)
     .subscribe(
       // this object contains our observer arguments
       // next for success, error for failure, complete for all
+
+      // Response was successful, so we're going to get all employees again and set all errors to false
     {
       next: data => {
         this.employees = [];
@@ -43,27 +56,26 @@ export class EmployeesComponent {
           this.getAllEmployees();
           this.officeDoesNotExist = false;
           this.maxEmployees = false;
-          console.log('success!');
         }
       },
-      error: error => { // some lambda for an error response
 
-        console.log(error.headers);
+      // Response was an error, so we're going to set the appropriate error to true
+      error: error => {
+
+        // If the error is a 400 and the header is badOfficeId, we set officeDoesNotExist to true
         if (error.status == 400 && error.headers.get('error') == 'badOfficeId') {
           this.officeDoesNotExist = true;
+
+        // If the error is a 400 and the header is maxEmployees, we set maxEmployees to true
         } else if (error.status == 400 && error.headers.get('error') == 'maxEmployees') {
           this.maxEmployees = true;
         }
-        this.getAllEmployees();
       },
-      // a lambda for something to do AFTER a successful response
-      // useful for void return HTTP methods like DELETE
-      complete: () => {
-        console.log('Complete');
-      }
     });
   }
 
+
+  // Method to get all employees and add them to the employees array
   getAllEmployees() {
     this.employees = [];
 
@@ -72,13 +84,14 @@ export class EmployeesComponent {
           let body: any = response.body;
 
           for (let item of body) {
-            // we're creating a new Department object for each item in the response
-            // and pushing it into our departments array
+            // we're creating a new Employee object for each item in the response
+            // and pushing it into our employee array
             this.employees.push(new Employee(item.employeeId, item.employeeFirstName, item.employeeLastName, item.employeeAddress, item.employeeSsn, item.employeeManagerId, item.office));
           }
         });
     }
 
+  // Method to get all employees sorted by name and add them to the employees array
   getEmployeesSortedByName() {
     this.employees = [];
 
@@ -87,48 +100,47 @@ export class EmployeesComponent {
           let body: any = response.body;
 
           for (let item of body) {
-            // we're creating a new Department object for each item in the response
-            // and pushing it into our departments array
-            console.log(item);
             this.employees.push(new Employee(item.employeeId, item.employeeFirstName, item.employeeLastName, item.employeeAddress, item.employeeSsn, item.employeeManagerId, item.office));
           }
         });
   }
 
+  // Get a specific employee by ID, used in the search feature. Will only show one employee card when used
   getEmployeeById() {
+
+    // Send the HTTP Request
     this.httpService.getEmployeeById(this.searchId).subscribe(
-      // this object contains our observer arguments
-      // next for success, error for failure, complete for all
     {
+      // response was successful, clear the employees array and add the employee to it
       next: data => {
         this.employees = [];
         if (data.body && data.body !== null) {
           this.employees.push(data.body as Employee);
           this.employeeDoesNotExist = false;
-          console.log('success!');
         }
       },
-      error: error => { // some lambda for an error response
-        console.log(error);
+      // response was an error, set the employeeDoesNotExist flag to true
+      error: error => {
         this.employeeDoesNotExist = true;
         this.getAllEmployees();
       },
-      // a lambda for something to do AFTER a successful response
-      // useful for void return HTTP methods like DELETE
-      complete: () => {
-        console.log('Complete');
-      }
     });
   }
 
+  // Method to delete an employee by ID
+  // This method will call the deleteEmployee method in the httpService
   deleteEmployee(employeeId: number) {
-    this.httpService.deleteEmployee(employeeId)
+    // Confirm the user wants to delete the employee
+    if (confirm("Are you sure you want to delete this employee? (ID: " + employeeId + ")")) {
+      this.httpService.deleteEmployee(employeeId)
       .subscribe(data => {
         this.getAllEmployees();
       });
+    }
   }
 
-  // this method runs when the deleteDepartmentEvent is emitted from the child component
+  // Listener for the delete event from the employee card component
+  // Pass the id into deleteEmployee()
   processDeleteEvent(employeeId: number) {
     this.deleteEmployee(employeeId);
   }
